Updating Makefile and nEvents Calculation

+++ Tools/condor/goNEvts.sh
@@ -15,8 +15,12 @@ export PATH=${PATH}:/cvmfs/cms.cern.ch/common
 export CMS_PATH=/cvmfs/cms.cern.ch
 
 #move the files from /home to /var/tmp
-mv ${_PWD}/$2.tar.gz ${_CONDOR_SCRATCH_DIR}
-mv ${_PWD}/TT.tar.gz ${_CONDOR_SCRATCH_DIR}
+if [ ! -f ${_CONDOR_SCRATCH_DIR}/$2.tar.gz ]; then
+    mv ${_PWD}/$2.tar.gz ${_CONDOR_SCRATCH_DIR}
+fi
+if [ ! -f ${_CONDOR_SCRATCH_DIR}/TT.tar.gz ]; then
+    mv ${_PWD}/TT.tar.gz ${_CONDOR_SCRATCH_DIR}
+fi
 
 cd ${_CONDOR_SCRATCH_DIR}
 
@@ -38,10 +42,9 @@ mv samplesModule.so ../obj
 pwd
 ls -lhrt
 
-python nEvts.py -s sampleSets.cfg -d "^$1$" > output_$1.txt
+python nEvts.py -s $3 -d "^$1$" > output_$1.txt
 ls -lhrt
 
 mv output_$1.txt ${_CONDOR_SCRATCH_DIR}
 
-rm $(echo $6 | sed 's|.*/||')
 rm -r ${_CONDOR_SCRATCH_DIR}/$2
